+++ packages/gatsby-plugin-orion-edit/src/utils/get-parent-path.js
@@ -1,0 +1,10 @@
+const getParentPath = ancestry => {
+  const directParent = ancestry.filter(ancestor => ancestor.direct)
+  if (directParent.length > 0) {
+    return directParent[0].ancestor.path
+  }
+
+  return ''
+}
+
+export default getParentPath