+++ packages/gatsby-plugin-orion-admin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.47.0](https://github.com/nearform/orion/compare/v2.46.1...v2.47.0) (2020-04-24)
+
+**Note:** Version bump only for package gatsby-plugin-orion-admin
+
+
+
+
+
+## [2.46.1](https://github.com/nearform/orion/compare/v2.46.0...v2.46.1) (2020-04-21)
+
+**Note:** Version bump only for package gatsby-plugin-orion-admin
+
+
+
+
+
 # [2.41.0](https://github.com/nearform/orion/compare/v2.40.0...v2.41.0) (2020-04-03)
 
 **Note:** Version bump only for package gatsby-plugin-orion-admin